requirements.txt
python-dotenv==1.0.1
customtkinter==5.2.2
pytesseract==0.3.13
easyocr==1.7.2
google-generativeai==0.8.3
Pillow==11.0.0
requests==2.32.3
keyboard==0.13.5
torch==2.5.1
winocr==0.0.15
numpy>=1.24.0
opencv-python>=4.10.0.84
pywin32==306
protobuf~=4.25.5
retrying==1.3.4
pyperclip==1.9.0
CTkMessagebox==2.5
aiohttp==3.9.3